[#2297] Move config models to separate files, fix type checking
