+++ docs/configuration/kic.rst
@@ -30,6 +30,7 @@ All settings:
     KIC_CONFIG_CONTACTMOMENTEN_SERVICE_API_ROOT
     KIC_CONFIG_CONTACTMOMENTEN_SERVICE_CLIENT_ID
     KIC_CONFIG_CONTACTMOMENTEN_SERVICE_SECRET
+    KIC_CONFIG_ENABLE
     KIC_CONFIG_KLANTEN_SERVICE_API_ROOT
     KIC_CONFIG_KLANTEN_SERVICE_CLIENT_ID
     KIC_CONFIG_KLANTEN_SERVICE_SECRET
