+++ docs/configuration/siteconfig.rst
@@ -36,6 +36,7 @@ All settings:
     SITE_EHERKENNING_ENABLED
     SITE_EMAIL_NEW_MESSAGE
     SITE_EMAIL_VERIFICATION_REQUIRED
+    SITE_ENABLE
     SITE_FOOTER_LOGO_TITLE
     SITE_FOOTER_LOGO_URL
     SITE_GA_CODE
