+++ docs/configuration/zgw.rst
@@ -40,6 +40,7 @@ All settings:
     ZGW_CONFIG_DOCUMENT_SERVICE_API_ROOT
     ZGW_CONFIG_DOCUMENT_SERVICE_CLIENT_ID
     ZGW_CONFIG_DOCUMENT_SERVICE_SECRET
+    ZGW_CONFIG_ENABLE
     ZGW_CONFIG_ENABLE_CATEGORIES_FILTERING_WITH_ZAKEN
     ZGW_CONFIG_FETCH_EHERKENNING_ZAKEN_WITH_RSIN
     ZGW_CONFIG_FORM_SERVICE_API_ROOT
