--- src/open_inwoner/configurations/bootstrap/typing.py
@@ -1,17 +1,0 @@
-from typing import TypeAlias
-
-from .models import (
-    DigiDOIDCConfigurationSettings,
-    KICConfigurationSettings,
-    SiteConfigurationSettings,
-    ZGWConfigurationSettings,
-    eHerkenningDOIDCConfigurationSettings,
-)
-
-ConfigSetting: TypeAlias = (
-    SiteConfigurationSettings
-    | KICConfigurationSettings
-    | ZGWConfigurationSettings
-    | DigiDOIDCConfigurationSettings
-    | eHerkenningDOIDCConfigurationSettings
-)
